apps/arcs/omega-reason-stream/requirements.txt
nats-py==2.8.0
